Adding notificationReducer to redux
