+++ packages/bierzo-wallet/src/store/index.ts
@@ -1,14 +1,17 @@
 import { applyMiddleware, combineReducers, compose, createStore, Middleware, Reducer, Store } from 'redux';
 import { StateType } from 'typesafe-actions';
 import { extensionReducer, ExtensionState } from './extension';
+import { notificationReducer, NotificationState } from './notifications';
 
 export interface RootReducer {
   extension: ExtensionState;
+  notifications: NotificationState;
 }
 
 const createRootReducer = (): Reducer<RootReducer> =>
   combineReducers({
     extension: extensionReducer,
+    notifications: notificationReducer,
   });
 
 export type RootState = StateType<ReturnType<typeof createRootReducer>>;
