+++ packages/bierzo-wallet/src/store/notifications/actions.ts
@@ -1,7 +1,15 @@
-import { ProcessedTx } from './state';
-import { AddPendingTransactionActionType } from './reducer';
+import { Action } from 'redux';
+import { ActionType } from 'typesafe-actions';
+import { ProcessedTx } from './reducer';
+
+export interface AddPendingTransactionActionType extends Action {
+  type: '@@notifications/ADD_PENDING_TRANSACTION';
+  payload: ProcessedTx;
+}
 
 export const addPendingTransactionAction = (transaction: ProcessedTx): AddPendingTransactionActionType => ({
   type: '@@notifications/ADD_PENDING_TRANSACTION',
   payload: transaction,
 });
+
+export type NotificationActions = ActionType<typeof addPendingTransactionAction>;
