+++ packages/bierzo-wallet/src/store/notifications/index.ts
@@ -1,0 +1,2 @@
+export * from './actions';
+export * from './reducer';
